backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.10.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
openai>=1.0.0
apscheduler>=3.10.0
